+++ js/data.js
@@ -1,4 +1,4 @@
-import {getRandomElement, getRandomIntInclusive} from './utils.js';
+
 const NAMES = [
   'Андрей',
   'Игорь',
@@ -35,22 +35,4 @@ const DESCRIPTION = [
 
 const PHOTO_NUMBER_ELEMENTS = 25;
 
-//функция для создания photo
-const createPhoto = () => ({
-
-  id: getRandomIntInclusive(1, 25),
-  url: `photos/${getRandomIntInclusive(1, 25)}.jpg`,
-  description: getRandomElement(DESCRIPTION),
-  likes: getRandomIntInclusive(15, 200),
-  comments: {
-    id: getRandomIntInclusive(1, 200),
-    avatar:` img/avatar-${getRandomIntInclusive(0, 6)}.svg`,
-    message: getRandomElement(MESSAGES),
-    name:getRandomElement(NAMES),
-  },
-});
-
-
-//СОЗДАНИЕ МАССИВА ИЗ N ЭЛЕМЕНТОВ
-const photo = Array.from({length:PHOTO_NUMBER_ELEMENTS}, createPhoto);
-console.log(photo);
+export {NAMES, MESSAGES, DESCRIPTION, PHOTO_NUMBER_ELEMENTS};
