+++ js/main.js
@@ -1,2 +1,3 @@
-import './data.js';
-
+import './createPhoto.js';
+import { photo } from './createPhoto.js';
+console.log(photo)
